requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
distlib==0.3.9
Django==5.1.2
django-ckeditor==6.7.2
django-import-export==4.3.1
django-js-asset==2.2.0
djangorestframework==3.15.2
filelock==3.16.1
pillow==11.0.0
platformdirs==4.3.6
sqlparse==0.5.1
tablib==3.7.0
tzdata==2024.2
virtualenv==20.27.0
